Add new controller for editing a sites transition date
We want to add a form to edit the details of a site. However, we already have a
form for editing a site's transition date.

This moves the action into its own controller, so that we can use the `edit`
action of the `SitesController` for the site details instead.
Add custom route for site dates

We want to use a new controller for site dates in order to keep our controllers
simple (http://jeromedalbert.com/how-dhh-organizes-his-rails-controllers/).
