+++ app/controllers/sites_controller.rb
@@ -3,10 +3,9 @@ class SitesController < ApplicationController
 
   include PaperTrail::Rails::Controller
 
-  before_action :find_site, only: %i[edit update show confirm_destroy destroy]
+  before_action :find_site, except: %i[new create]
   before_action :find_organisation, only: %i[new create]
-  before_action :check_user_is_gds_editor, only: %i[edit update]
-  before_action :check_user_is_site_manager, only: %i[new create confirm_destroy destroy]
+  before_action :check_user_is_site_manager, except: %i[show]
 
   def new
     @site_form = SiteForm.new(organisation_slug: @organisation.whitehall_slug)
@@ -19,16 +18,6 @@ class SitesController < ApplicationController
       redirect_to site_path(site), flash: { success: "Transition site created" }
     else
       render :new
-    end
-  end
-
-  def edit; end
-
-  def update
-    if @site.update(update_params)
-      redirect_to site_path(@site), flash: { success: "Transition date updated" }
-    else
-      redirect_to edit_site_path(@site), flash: { alert: "We couldn't save your change" }
     end
   end
 
@@ -81,19 +70,8 @@ private
     )
   end
 
-  def update_params
-    params.require(:site).permit(:launch_date)
-  end
-
   def destroy_params
     params.require(:delete_site_form).permit(:abbr_confirmation)
-  end
-
-  def check_user_is_gds_editor
-    unless current_user.gds_editor?
-      message = "Only GDS Editors can access that."
-      redirect_to site_path(@site), alert: message
-    end
   end
 
   def check_user_is_site_manager
