+++ app/views/sites/show.html.erb
@@ -27,7 +27,7 @@
       <%= date_or_not_yet(@site.launch_date) %>
       <br />
       <% if current_user.gds_editor? || current_user.admin? %>
-        <span class="small"><%= link_to 'Edit date', edit_site_path(@site), class: 'link-inherit' %></span>
+        <span class="small"><%= link_to 'Edit date', site_edit_date_path(@site), class: 'link-inherit' %></span>
       <% end %>
     </div>
